+++ app/helpers/application_helper.rb
@@ -20,10 +20,6 @@ module ApplicationHelper
     'data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw=='
   end
 
-  def check_root
-    params[:controller].eql?('pages') and params[:action].eql?('home')
-  end
-
   def bootstrap_form_for(object, options = {}, &block)
     options[:builder] = BootstrapFormBuilder
     form_for(object, options, &block)
@@ -31,10 +27,6 @@ module ApplicationHelper
 
   def options_for_array_collection(model, attr_name, *args)
     options_for_select("#{model}::#{attr_name.to_s.upcase}".safe_constantize.map { |e| [ model.human_attribute_name("#{attr_name}_#{e}"), e] }, *args)
-  end
-
-  def collection_count(collection)
-    "(#{collection.size})" if collection.any?
   end
 
   def background_jobs_available?