+++ spec/models/feedback_spec.rb
@@ -1,0 +1,14 @@
+require 'spec_helper'
+
+describe Feedback do
+  describe '.fixed' do
+
+    it "returns true if the status is feedback is fixed" do
+      fixed_feedback = FactoryGirl.create :feedback, status: 'fixed'
+      open_feedback = FactoryGirl.create :feedback
+
+      expect(fixed_feedback.fixed?).to be_true
+      expect(open_feedback.fixed?).to be_false
+    end
+  end
+end